requirements.txt
DrissionPage>=4.0.0
psutil>=5.9.3
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
sqlalchemy>=2.0.0
asyncpg>=0.27.0
python-dotenv>=1.0.0
playwright==1.41.2
aiosqlite==0.21.0
fake-useragent==2.1.0
python-multipart
colorama>=0.4.6